cmd/db/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
geojson==2.5.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
pyowm==3.3.0
PySocks==1.7.1
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
Werkzeug==2.0.3